Add timestamp to jsonschema (#38)
Summary of the issue
Add-on metadata may accept a timestamp, to sort add-ons or perform other queries.

Development approach
Update the json schema to accept an optional "submissionTime", corresponding to a timestamp. We use milliseconds instead of seconds, in case similar values are added in the future, for a better comparison.
When creating a json file, the current time will be assigned to submissionTime.
Updated tests, removing the submissionTime before comparing actual and expected json files.