Remove items whose value is None before dumping to JSON (#53)
Some add-on submissions, such as this one are failing.

The problem appears to be that optional values in the JSON metadata are being given a value of null, rather than omitted. Since null is a value, and the datatype for these keys is string, the validation fails.

Thus, remove any None values from the add-on metadata before dumping to JSON. I believe this is safe, as we have already validated that all required keys are present, so any Nones on the dataclass are optional keys that haven't been provided.